Extract useMedicationsMap
