+++ app/(dashboard)/patients/clientUtils.ts
@@ -7,6 +7,8 @@
 //
 import { ObservationType } from '@/modules/firebase/utils'
 import { strategy } from '@/packages/design-system/src/utils/misc'
+import { useMemo } from 'react'
+import { MedicationsData } from '@/app/(dashboard)/patients/utils'
 
 export const getObservationTypeUnits = (observationType: ObservationType) =>
   strategy(
@@ -67,4 +69,14 @@ export const getUnitOfObservationType = (
   return newUnit
 }
 
-export type LabUnit = ReturnType<typeof getObservationTypeUnits>[number]
+export const useMedicationsMap = (
+  medications: MedicationsData['medications'],
+) =>
+  useMemo(() => {
+    const entries = medications.flatMap((medicationClass) =>
+      medicationClass.medications.map(
+        (medication) => [medication.id, medication] as const,
+      ),
+    )
+    return new Map(entries)
+  }, [medications])
